Make generate_QMfiles a method of QMCalculator
